Call ensure_paths in .prow.sh
There's a mismatch between the go version installed in the prow job
(1.17) and the version wanted to do the build (1.16), because of it
there's a new path in run_with_go that will attempt to download 1.16 but
fail because an env var is not set.

Calling ensure_paths makes sure that the env_var is set, this is a
change outside of release-tools.